add color, constant and loader modules
